Correcting bug on erroring submit of agreement
Prior to this patch, when you would enter SeniorThesis information,
checking the agreement but the server side validation failed, when the
form would be rendered the "I have read and accept..." checkbox would
be selected but the "Create Senior thesis" would be disabled. Then
clicking the checkbox would renable the button, but would trigger a
failed validation.