Add nginx route to handle 502 errors from killed unicorns.
As opposed to a non-existant unicorn. We use some care to keep
a redirect loop from happening.

This is intended to handle requests that are uploading large files
and take too long. Typically Fedora takes a while to ingest files,
and if it takes more than 60s the unicorn master kills process handling
the request. This configuration will catch that error and give the user
a nicer error message.