Add route to pass ids of objects to be indexed
The route acepts a JSON array of the object ids to reindex.
(either the full fedora PID or just the NOID portion).
The reindexing happens as a bunch of background jobs, each job handling
up to 10 ids each.

The route is behind a new constraint which requires an API key.
There is a config file for API keys, and it is added to the deploy.
(hence future deploys may break until the file is added to the secrets
repos).

Addresses DLTP-117