One-time-use URLs will be good for a period of time
Expiring a one-time-use token after a single request is not a
sufficiently sophisticated solution. For example, Chrome uses range
requests to incrementally fetch PDFs. The prior implementation of this
feature would deny all rage requests after the first request resulting
in Chrome rendering a "empty" PDF.

@dbrower suggested making the URLs valid for a period of time, like a
day. This should allow patrons sufficient time to access the desired
materials.

Alternatively the permission to access the desired file could be stored
in the application session or a different temporary cookie.