formInputMask widget documentation; conditional field display
Instructions on initialization and clarifying comments on intent.

Form fields with a value that would normally be hidden by an input mask
are shown anyway. Applying an input mask does not clear form field
values so. Before this change it would be possible to submit an
attribute that will be persisted and displayed that is not shown on the
form. This is confusing.

Changed conditional formatting as recommended by jshint.

The event object passed to toggleTypedInputFieldsHandler is ALREADY a
jQuery object. Re-casting it caused `val` function to return `undefined`.