Test cases for inline and block level rich text rendering
"Inline" text rendering is for anything that can be captured by an
`input` field and the output would be contained in a single `<p>` tag.
This is the most common the use cases for rich text in CurateND. It
includes object attributes and titles. In these contexts the list style
syntax causes undesired behavior.

"Block" text rendering is for anything that can be captured by a
`textarea` field and the output would be contained in multiple `<p>`
tags. It is used sparingly in CurateND.