Let Disadis set the headers
This line had previously just been `content_options`. This did not alter
any behavior because `content_options` returns a hash literal.

`send_file_headers! content_options` is a convention used in several
places in the download behavior. However, when it is included here it
causes a browser error because the disposition of the file is set twice.