Adding a preliminary library collection model
I want to work on the specifics of the Library Collection model instead
of attempting to deprecate some of the existing Collection model
behavior.

At a future point, we can either migrate the data from
LibraryCollection to Collection or obliterate the existing Collection
model when we have enough of the given behavior worked out.