Resequence Document attributes
Document attributes were not in the same sequence in the two files,
which made it more difficult to locate and/or validate them. Attributes
are now in the sequence they display on the input screen, with all
common attributes first, followed by book-only attributes. Apparently
unused attributes were left intact, but moved to the end of the files.