Removing writing logs to unicorn stdout log
The following line shows why we are likely writing $stderr to unicorn
https://bogomips.org/unicorn.git/tree/lib/unicorn.rb?h=v4.6.3#n76

As I was debugging this I encountered a caching-type behavior. If I
changed the code on how unicorn starts then deployed the changes, the
restart unicorn wouldn't always use the right RAILS_ENV.

What I needed to do to get this working was to SSH onto the machine and
stop the unicorn process then restart it. It was as thought the -E
option were either cached or continued to be used.