Display dates for patents and senior theses
https://jira.library.nd.edu/browse/DLTP-1401

Quick summary of work vs issue:

1) senior theses are not indexing the date created. I can't figure out
why and it's not worth a lot of time since we don't add these any more.
So I changed the partial to show the date created from the model rather
than from the solr document (we already had that info in the partial so
it's no extra time to load it).
2) if finding aids had the date created, it would already be showing, so
this is beyond the scope of this issue.
3) Images already show date created if it exists.
4) Added issued date for patents.