Multi Upload of files after creation
DLTP-1341, Restructured the file creation to allow for multiple files to be uploaded at once. Adjusted the input options to reflect.

Minor fix

needed to for variable to array to fix QA failure

Specs additions

Updates scepcs that needed changed for multi file upload and added a spec to specifically for the multi file upload process in generic_file_actor