Ensuring expanded ToS allows acceptance
Prior to this commit, the HTML modal was rendering inside the form.
This would work, except that the modal JS would pull that HTML
and render the DOM elements outside of the form. This resulted in a
loss of the form context, and thus the submit buttons didn't know
where to submit the accept or decline action.

By moving the modal content outside of the form, and adding a form
around the submit buttons:

* the expected behavior is restored
* we don't violate HTML standards by having a form nested within
  another form

In addition, I've added a confirmation flash message to let the end
user know that the ToS was accepted.

DLTP-1521