Restructure representative partials
Just a little cleanup...
* Added comments to clarify the sections used to display representatives.
* Removed unnecessary code from lower-level audio & image partials.
* Moved code from curate_helper into partial since it was only accessed from one place.

I would have liked to clean this up a bit more, but the representative_media partial
was also accessed directly for library collections, so some of the code couldn't be removed.

I also would have liked to limit the javascript spinner to only cases where we might have a
manifest, but that change would require a pretty large restructure of these partials.