Removing asset debugging
This commit was a bit of a stab in the dark. Prior to the commit, the
staging environment's compiled asset declaration were as follows:

```html
<script src="/assets/modernizr.js"></script>
<script src="/assets/application.js"></script>
```

However, since we compiled assets on the server, those files did not
exist. (Instead we had a compiled file in ./public/assets called "application-92dee9155d309d27b1416c3ecc4001b9c425065934e5bc363c363e0a66edeb14.js")

The rendered HTML after this change is:

```html
<script src="/assets/modernizr-302af516043e5636aa627ea0f13080021c3b74975a20c1850a9acbf1512854ce.js"></script>
<script src="/assets/application-92dee9155d309d27b1416c3ecc4001b9c425065934e5bc363c363e0a66edeb14.js"></script>
```

What appears to have happened is that from Rails 4.1 to 4.2 the behavior
of `config.assets.debug` shifted a bit. By saying `config.assets.debug = false`
the asset pipeline will compile the assets AND use the sha calculated
for the compiled as the URL.