Adding a documented, though undesired, behavior
As reported by Mikala, she encountered a situation where the title of
an object was something along the lines of "1. My Special Title". On the
show page for that object, things were fine. On the search results the
title showed as blank.

This spec demonstrates known but unexpected behavior. It is not that we
want this behavior, we just want to acknowledge this behavior.