Ignore IOError when requesting bendo to cache content
The content caching mechanism relies on not reading the entire body
of a GET request. Whereas the HTTP spec says the body of a GET request
should not be terminated early on purpose. Thus errors are expected.

Revise bendo to provide a different mechanism and remove this code when
that is in place.

DLTP-1787