Remove Contributor Institution From Deposit
Completes https://jira.library.nd.edu/browse/DLTP-1561

Contributor Institution is defined for work types: Article, Document, Image, Dataset.
No datasets have the field indexed in Solr.
There are 9 articles, 11 images, and 223 documents.

It is not included on any views for Article.
It is on add, edit, and show views for Document.
It is on add & edit views for Image and Dataset.

The request was to remove from deposit. Removing from add also removes
from edit because they share the same partials.

It was only showing for type Document, and it remains in that view.